fix(runtime,client): split the ledger guard along the package boundary — CI has no client dist in runtime's test task
Test Core failed on the first push: the conformance test imported the built
client (`../../client/dist/index.mjs`), assuming the dist exists whenever
tests run. Wrong assumption — turbo's per-package test tasks build only their
own dependency closure, and runtime deliberately has no edge to client (the
reverse edge exists, so adding one is a build cycle turbo rejects). In CI the
dist therefore never exists for runtime's suite.

The fix is structural, not a skip: each package now verifies its own half.

- runtime keeps the dispatcher directions (registry domain ↔ ledger, both
  ways) plus ledger hygiene and the gap ratchet — no client import at all.
- client gains `route-ledger-coverage.test.ts`, which instantiates the real
  ObjectStackClient and asserts every ledger-named method resolves to a
  function. The ledger is imported as a relative SOURCE file: it is pure data
  with zero imports, and it lives in runtime because that is where routes are
  declared.
- the now-unneeded @objectstack/core/logger vitest alias is reverted.

Skipping when the dist is missing was rejected: Test Core would never
exercise the client direction, and the guard's whole point is that this
failure mode cannot go quiet again.

Verified: phantom-method negative check bites in the new home; client suite
117 passed, runtime suite 653 passed; eslint clean.
